+++ HISTORY.rst
@@ -11,3 +11,14 @@ History
 0.0.6 (2021-05-30)
 ------------------
 * Basic set of Text & Number matchers
+
+1.0.0 (2021-05-30)
+------------------
+* Better testing and refactoring in the spirit of DRY
+
+1.1.0 (2021-05-30)
+------------------
+* Adding existence logic: None, NotNone, NoneOrEmpty, NotNoneOrEmpty
+
+
+